refactor: modularize FileBrowser and info modal
