+++ src/modules/filemanager/components/FileBrowser/FileBrowser.scss
@@ -7,7 +7,6 @@
   position: relative;
 }
 
-/* Content area */
 .fm-file-browser-content {
   position: relative;
   overflow: auto;
@@ -16,7 +15,6 @@
   flex: 1 1 auto;
 }
 
-/* Header grid + responsive templates for search/non-search */
 .fm-file-browser-content-header {
   display: grid;
   padding: 12px;
@@ -25,17 +23,14 @@
   border-bottom: 1px solid rgb(226, 232, 240);
 }
 
-/* Normal: ☐ | Name | Size | Date */
 .fm-file-browser-content[data-search-mode='false'] .fm-file-browser-content-header {
   grid-template-columns: 32px 2fr 1fr 1fr;
 }
 
-/* Search: ☐ | Name | Drive | Size | Date */
 .fm-file-browser-content[data-search-mode='true'] .fm-file-browser-content-header {
   grid-template-columns: 32px 2fr 1.1fr 1fr 1fr;
 }
 
-/* Header cells */
 .fm-file-browser-content-header-item {
   display: flex;
   align-items: center;
@@ -47,11 +42,10 @@
     margin: 0;
     width: 16px;
     height: 16px;
-    accent-color: rgb(237, 129, 49); /* orange */
+    accent-color: rgb(237, 129, 49);
   }
 }
 
-/* Center the header checkbox like rows */
 .fm-file-browser-content-header-item.fm-checkbox {
   display: flex;
   align-items: center;
@@ -61,7 +55,7 @@
 .fm-file-browser-content-header-item-icon {
   width: 16px;
   height: 16px;
-  color: rgb(237, 129, 49); /* orange */
+  color: rgb(237, 129, 49);
 }
 
 .fm-file-browser-content-body {
@@ -69,7 +63,6 @@
   position: relative;
 }
 
-/* Footer */
 .fm-file-browser-footer {
   display: flex;
   align-items: center;
@@ -80,14 +73,26 @@
   background-color: #ededed;
 }
 
-.fm-file-browser-footer > * { flex: 0 0 auto; }
-.fm-file-browser-footer > :nth-child(1) { margin-right: 8px; }
-.fm-file-browser-footer > :nth-child(3) { margin-left: auto; }
-.fm-file-browser-footer { column-gap: 12px; }
+.fm-file-browser-footer > * {
+  flex: 0 0 auto;
+}
+.fm-file-browser-footer > :nth-child(1) {
+  margin-right: 8px;
+}
+.fm-file-browser-footer > :nth-child(3) {
+  margin-left: auto;
+}
+.fm-file-browser-footer {
+  column-gap: 12px;
+}
 
-/* Context menu */
-.fm-file-browser-context-menu { position: absolute; z-index: 1000; }
-.fm-file-browser-context-menu[data-drop='up'] { transform-origin: bottom left; }
+.fm-file-browser-context-menu {
+  position: absolute;
+  z-index: 1000;
+}
+.fm-file-browser-context-menu[data-drop='up'] {
+  transform-origin: bottom left;
+}
 .fm-file-browser-context-menu[data-drop='up'] .caret {
   transform: rotate(180deg);
   bottom: -6px;
@@ -98,15 +103,23 @@
   margin: 4px;
   padding: 8px;
 
-  &:hover { background-color: #d1d1d1; cursor: pointer; }
-  &.red { color: rgb(220, 38, 38); }
+  &:hover {
+    background-color: #d1d1d1;
+    cursor: pointer;
+  }
+  &.red {
+    color: rgb(220, 38, 38);
+  }
 }
 
-.fm-context-item-border { border-bottom: 1px solid #d1d1d1; }
+.fm-context-item-border {
+  border-bottom: 1px solid #d1d1d1;
+}
 
-.fm-upload-download-indicator { text-decoration: underline; }
+.fm-upload-download-indicator {
+  text-decoration: underline;
+}
 
-/* Drag overlay */
 .fm-drag-overlay {
   position: absolute;
   inset: 0;
@@ -131,4 +144,7 @@
   border: 1px solid rgba(0, 0, 0, 0.06);
 }
 
-.fm-drop-hint { padding: 24px; color: #6b7280; }
+.fm-drop-hint {
+  padding: 24px;
+  color: #6b7280;
+}
