+++ src/modules/filemanager/components/FileBrowser/FileItem/FileItem.tsx
@@ -9,8 +9,8 @@ import { GetInfoModal } from '../../GetInfoModal/GetInfoModal'
 import { VersionHistoryModal } from '../../VersionHistoryModal/VersionHistoryModal'
 import { DeleteFileModal } from '../../DeleteFileModal/DeleteFileModal'
 import { RenameFileModal } from '../../RenameFileModal/RenameFileModal'
-import { buildGetInfoGroups } from '../../GetInfoModal/buildFileInfoGroups'
-import type { FilePropertyGroup } from '../../GetInfoModal/buildFileInfoGroups'
+import { buildGetInfoGroups } from '../../../utils/infoGroups'
+import type { FilePropertyGroup } from '../../../utils/infoGroups'
 import { useView } from '../../../providers/FMFileViewContext'
 import type { DriveInfo, FileInfo } from '@solarpunkltd/file-manager-lib'
 import { useFM } from '../../../providers/FMContext'
